re-structured Shaope component
