+++ src/foundations/shape/index.ts
@@ -1,6 +1,7 @@
 import Shape from './Shape';
 
 export * from './Shape.types';
+export * from './Shape.props';
 export * from './Shape';
 
 export default Shape;
